+++ ledger/api/character/template.py
@@ -15,7 +15,6 @@ from ledger.hooks import get_extension_logger
 logger = get_extension_logger(__name__)
 
 
-# pylint: disable=too-many-locals, too-many-branches, too-many-statements
 class LedgerTemplateApiEndpoints:
     tags = ["CharacerLedgerTemplate"]